Add Application class
